# specs/001-build-football-admin-app/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tournament Administration API (Phase One)
  version: 0.1.0
  description: >-
    REST contract for the modern football tournament administration application. Covers admin,
    team manager, and public scoreboard surfaces required for phase one delivery.
servers:
  - url: https://api.tournament.local
    description: Local development
...
paths:
  /api/auth/invitations:
    post:
      tags: [auth]
      summary: Invite a new organizer or team manager
      operationId: create_invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400':
          $ref: '#/components/responses/ProblemDetails'
  /api/competitions:
    post:
      tags: [competitions]
      summary: Create a new competition and initial edition
      operationId: create_competition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompetitionRequest'
      responses:
        '201':
          description: Competition and draft edition created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionWithEdition'
        '422':
          $ref: '#/components/responses/ProblemDetails'
  /api/competitions/{competition_id}:
    get:
      tags: [competitions]
      summary: Fetch competition with editions and roles
      operationId: get_competition
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
      responses:
        '200':
          description: Competition detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionDetail'
        '403':
          $ref: '#/components/responses/ProblemDetails'
        '404':
          $ref: '#/components/responses/ProblemDetails'
...
  /api/editions/{edition_id}:
    get:
      tags: [competitions]
      summary: Fetch edition scoreboard configuration
      operationId: get_edition_scoreboard
      parameters:
        - $ref: '#/components/parameters/EditionId'
      responses:
        '200':
          description: Edition scoreboard configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditionScoreboardView'
        '403':
          $ref: '#/components/responses/ProblemDetails'
    patch:
      tags: [competitions]
      summary: Update edition settings
      operationId: update_edition
      parameters:
        - $ref: '#/components/parameters/EditionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEditionRequest'
      responses:
        '200':
          description: Edition updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditionScoreboardView'
        '403':
          $ref: '#/components/responses/ProblemDetails'
  /api/editions/{edition_id}/stages:
    post:
      tags: [competitions]
      summary: Create a stage (group or knockout)
      operationId: create_stage
      parameters:
        - $ref: '#/components/parameters/EditionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStageRequest'
      responses:
        '201':
          description: Stage created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stage'
        '400':
          $ref: '#/components/responses/ProblemDetails'
...
  /api/editions/{edition_id}/scoreboard/highlights:
    post:
      tags: [competitions]
      summary: Trigger a scoreboard highlight overlay
      operationId: trigger_scoreboard_highlight
      parameters:
        - $ref: '#/components/parameters/EditionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerScoreboardHighlightRequest'
      responses:
        '201':
          description: Highlight activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditionScoreboardView'
        '400':
          $ref: '#/components/responses/ProblemDetails'
        '403':
          $ref: '#/components/responses/ProblemDetails'
        '404':
          $ref: '#/components/responses/ProblemDetails'
    delete:
      tags: [competitions]
      summary: Clear the active scoreboard highlight
      operationId: clear_scoreboard_highlight
      parameters:
        - $ref: '#/components/parameters/EditionId'
      responses:
        '200':
          description: Highlight cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditionScoreboardView'
        '403':
          $ref: '#/components/responses/ProblemDetails'
        '404':
          $ref: '#/components/responses/ProblemDetails'
...
components:
  parameters:
    CompetitionId:
      name: competition_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EditionId:
      name: edition_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ProblemDetails:
      description: Problem Details error payload
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Invitation:
      type: object
      required: [id, email, role, scope]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [global_admin, competition_admin, team_manager]
        scope:
          type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [global, competition, edition, team]
            id:
              type: string
              format: uuid
              nullable: true
        expires_at:
          type: string
          format: date-time
    CreateInvitationRequest:
      type: object
      required: [email, role, scope]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [global_admin, competition_admin, team_manager]
        scope:
          $ref: '#/components/schemas/InvitationScope'
        message:
          type: string
          maxLength: 500
    InvitationScope:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [global, competition, edition, team]
        id:
          type: string
          format: uuid
          nullable: true
    CompetitionWithEdition:
      type: object
      required: [competition, edition]
      properties:
        competition:
          $ref: '#/components/schemas/Competition'
        edition:
          $ref: '#/components/schemas/Edition'
    CompetitionDetail:
      allOf:
        - $ref: '#/components/schemas/Competition'
        - type: object
          properties:
            editions:
              type: array
              items:
                $ref: '#/components/schemas/Edition'
            administrators:
              type: array
              items:
                $ref: '#/components/schemas/UserSummary'
    Competition:
      type: object
      required: [id, name, slug, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        primary_venue_id:
          type: string
          format: uuid
          nullable: true
        scoreboard_theme:
          $ref: '#/components/schemas/ScoreboardTheme'
        created_at:
          type: string
          format: date-time
    CreateCompetitionRequest:
      type: object
      required: [name, slug, default_edition]
      properties:
        name:
          type: string
        slug:
          type: string
        default_edition:
          $ref: '#/components/schemas/CreateEditionRequest'
    Edition:
      type: object
      required: [id, competition_id, label, slug, status, registration_window]
      properties:
        id:
          type: string
          format: uuid
        competition_id:
          type: string
          format: uuid
        label:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        format:
          type: string
          enum: [round_robin, knockout, hybrid]
        registration_window:
          $ref: '#/components/schemas/RegistrationWindow'
        scoreboard_rotation_seconds:
          type: integer
          minimum: 2
          default: 5
        scoreboard_theme:
          $ref: '#/components/schemas/ScoreboardTheme'
        published_at:
          type: string
          format: date-time
          nullable: true
    CreateEditionRequest:
      type: object
      required: [label, slug, format, registration_window]
      properties:
        label:
          type: string
        slug:
          type: string
        format:
          type: string
          enum: [round_robin, knockout, hybrid]
        registration_window:
          $ref: '#/components/schemas/RegistrationWindow'
        scoreboard_theme:
          $ref: '#/components/schemas/ScoreboardTheme'
    UpdateEditionRequest:
      type: object
      properties:
        label:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        scoreboard_rotation_seconds:
          type: integer
          minimum: 2
        scoreboard_theme:
          $ref: '#/components/schemas/ScoreboardTheme'
    ScoreboardTheme:
      type: object
      properties:
        primary_color:
          type: string
          pattern: '^#([A-Fa-f0-9]{6})$'
        secondary_color:
          type: string
          pattern: '^#([A-Fa-f0-9]{6})$'
        background_image_url:
          type: string
          format: uri
          nullable: true
    EditionScoreboardHighlight:
      type: object
      required: [message, expires_at, remaining_seconds]
      properties:
        message:
          type: string
        expires_at:
          type: string
          format: date-time
        remaining_seconds:
          type: integer
          minimum: 0
    EditionScoreboardView:
      type: object
      required: [edition, highlight]
      properties:
        edition:
          $ref: '#/components/schemas/Edition'
        highlight:
          $ref: '#/components/schemas/EditionScoreboardHighlight'
          nullable: true
    TriggerScoreboardHighlightRequest:
      type: object
      required: [message, duration_seconds]
      properties:
        message:
          type: string
          maxLength: 160
        duration_seconds:
          type: integer
          minimum: 5
          maximum: 600
    RegistrationWindow:
      type: object
      required: [opens_at, closes_at]
      properties:
        opens_at:
          type: string
          format: date-time
        closes_at:
          type: string
          format: date-time
    Stage:
      type: object
      required: [id, edition_id, name, stage_type]
      properties:
        id:
          type: string
          format: uuid
        edition_id:
          type: string
          format: uuid
        name:
          type: string
        stage_type:
          type: string
          enum: [group, bracket]
        order:
          type: integer
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    CreateStageRequest:
      type: object
      required: [name, stage_type]
      properties:
        name:
          type: string
        stage_type:
          type: string
          enum: [group, bracket]
        groups:
          type: array
          description: Required for group stages
          items:
            $ref: '#/components/schemas/CreateGroupRequest'
    Group:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        standings:
          type: array
          items:
            $ref: '#/components/schemas/Standing'
    CreateGroupRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
...
    UserSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
...
    Standing:
      type: object
      required: [entry_id, played, won, drawn, lost, goals_for, goals_against, points]
      properties:
        entry_id:
          type: string
          format: uuid
        position:
          type: integer
        played:
          type: integer
        won:
          type: integer
        drawn:
          type: integer
        lost:
          type: integer
        goals_for:
          type: integer
        goals_against:
          type: integer
        goal_difference:
          type: integer
        points:
          type: integer
        fair_play_score:
          type: integer
          nullable: true
...